Represent graphics with a scene graph
We build on top of the Skia graphics library (https://skia.org/)
and its Python bindings (https://kyamagu.github.io/skia-python/).
We still rely on Pillow for displaying images across OSes and
creating GIFs.

This decouples the creation of graphics and their rendering,
and avoids painful workarounds that were previously needed (in
particular for rotation around the pinning position).