Use 4x SSAA (supersampling) to obtain antialiasing
Analytical methods create seams or other unwanted behaviors for our case of composition geometry.
See https://github.com/LuCEresearchlab/pytamaro/issues/34

We can afford to render the graphic on a surface that is 4x bigger (twice the width and twice the height) and then resample the graphic with a linear filter.

Skia's Analytical AntiAliasing is still enabled for text, as that leads to a nicer image and text is normally not composed with adjacent shapes.