judicious-docs: add examples and generate output programmatically
Add a curated set of examples.
The "transformer" script localizes them using the language-specific API and generates the corresponding output.
This data extends what Sphinx produces.

(Having the examples inside the docstring would have been better, but it is almost impossible to use RST to encode nice-looking examples for the HTML builder *and* easily extract that data for Judicious.)