Error message for graphics that are too big to show
Skia doesn't support surfaces/images > 2GB in size (https://skia-review.googlesource.com/c/skia/+/471764). When attempting to create such a big image, skia-python fails with an uninformative error message. This commit introduces a properly-localized, learner-friendly message.

Anti-aliasing supersampling is adjusted to avoid creating images that go beyond that limit only because of supersampling. We take this as an opportunity to increase the scaling factor from 2 to 3 (4x to 9x SSAA) when images are sufficiently small.